move ExitError from solver/errdefs to frontend/gateway/errdefs
